Updated: permissions for update module

+++ Http/Controllers/Backend/AttributeGroupsController.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
+use Illuminate\Support\Facades\Auth;
 use VaahCms\Modules\Store\Models\Attribute;
 use VaahCms\Modules\Store\Models\AttributeGroup;
 
@@ -20,12 +21,18 @@ class AttributeGroupsController extends Controller
 
     public function getAssets(Request $request)
     {
+        if (!Auth::user()->hasPermission('has-access-of-module-section')) {
+            $response['success'] = false;
+            $response['errors'][] = trans("vaahcms::messages.permission_denied");
+
+            return response()->json($response);
+        }
 
         try{
 
             $data = [];
 
-            $data['permission'] = [];
+            $data['permissions'] = \Auth::user()->permissions(true);
             $data['rows'] = config('vaahcms.per_page');
 
             $data['fillable']['columns'] = AttributeGroup::getFillableColumns();
@@ -78,6 +85,12 @@ class AttributeGroupsController extends Controller
     //----------------------------------------------------------
     public function updateList(Request $request)
     {
+        if (!Auth::user()->hasPermission('can-update-module')) {
+            $response['success'] = false;
+            $response['errors'][] = trans("vaahcms::messages.permission_denied");
+
+            return response()->json($response);
+        }
         try{
             return AttributeGroup::updateList($request);
         }catch (\Exception $e){
@@ -96,7 +109,12 @@ class AttributeGroupsController extends Controller
     //----------------------------------------------------------
     public function listAction(Request $request, $type)
     {
+        if (!Auth::user()->hasPermission('can-update-module')) {
+            $response['success'] = false;
+            $response['errors'][] = trans("vaahcms::messages.permission_denied");
 
+            return response()->json($response);
+        }
 
         try{
             return AttributeGroup::listAction($request, $type);
@@ -116,6 +134,12 @@ class AttributeGroupsController extends Controller
     //----------------------------------------------------------
     public function deleteList(Request $request)
     {
+        if (!Auth::user()->hasPermission('can-update-module')) {
+            $response['success'] = false;
+            $response['errors'][] = trans("vaahcms::messages.permission_denied");
+
+            return response()->json($response);
+        }
         try{
             return AttributeGroup::deleteList($request);
         }catch (\Exception $e){
@@ -133,6 +157,12 @@ class AttributeGroupsController extends Controller
     //----------------------------------------------------------
     public function fillItem(Request $request)
     {
+        if (!Auth::user()->hasPermission('can-update-module')) {
+            $response['success'] = false;
+            $response['errors'][] = trans("vaahcms::messages.permission_denied");
+
+            return response()->json($response);
+        }
         try{
             return AttributeGroup::fillItem($request);
         }catch (\Exception $e){
@@ -150,6 +180,12 @@ class AttributeGroupsController extends Controller
     //----------------------------------------------------------
     public function createItem(Request $request)
     {
+        if (!Auth::user()->hasPermission('can-update-module')) {
+            $response['success'] = false;
+            $response['errors'][] = trans("vaahcms::messages.permission_denied");
+
+            return response()->json($response);
+        }
         try{
             return AttributeGroup::createItem($request);
         }catch (\Exception $e){
@@ -184,6 +220,12 @@ class AttributeGroupsController extends Controller
     //----------------------------------------------------------
     public function updateItem(Request $request,$id)
     {
+        if (!Auth::user()->hasPermission('can-update-module')) {
+            $response['success'] = false;
+            $response['errors'][] = trans("vaahcms::messages.permission_denied");
+
+            return response()->json($response);
+        }
         try{
             return AttributeGroup::updateItem($request,$id);
         }catch (\Exception $e){
@@ -201,6 +243,12 @@ class AttributeGroupsController extends Controller
     //----------------------------------------------------------
     public function deleteItem(Request $request,$id)
     {
+        if (!Auth::user()->hasPermission('can-update-module')) {
+            $response['success'] = false;
+            $response['errors'][] = trans("vaahcms::messages.permission_denied");
+
+            return response()->json($response);
+        }
         try{
             return AttributeGroup::deleteItem($request,$id);
         }catch (\Exception $e){
@@ -218,6 +266,12 @@ class AttributeGroupsController extends Controller
     //----------------------------------------------------------
     public function itemAction(Request $request,$id,$action)
     {
+        if (!Auth::user()->hasPermission('can-update-module')) {
+            $response['success'] = false;
+            $response['errors'][] = trans("vaahcms::messages.permission_denied");
+
+            return response()->json($response);
+        }
         try{
             return AttributeGroup::itemAction($request,$id,$action);
         }catch (\Exception $e){
